requirements.txt
Django==5.0.6
psycopg2==2.9.9
pgvector==0.3
django-environ==0.11
openai==1.25
djangorestframework==3.15